Implemented Authorization with gates

+++ app/Http/Controllers/QuestionsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -33,7 +34,7 @@ class QuestionsController extends Controller
     {
         $question = new Question();
 
-        return view('questions.create', compact('questions'));
+        return view('questions.create', compact('question'));
     }
 
     /**
@@ -73,6 +74,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(Gate::denies('update-question', $question)){
+
+            abort(403, "Access is denied");
+        }
+        
         return view('questions.edit', compact('question'));
     }
 
@@ -85,6 +91,12 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+
+        if(Gate::denies('update-question', $question)){
+
+            abort(403, "Access is denied");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Your questions has been updated');
@@ -98,6 +110,12 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+
+        if(Gate::denies('delete-question', $question)){
+
+            abort(403, "Access is denied");
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Your question has been deleted');
